DRY out IP method code
